refactor: mock handler 분리 및 불필요 참조 삭제
